Move info components and ui components
