+++ src/components/info-components/styles/TransformerInfo.css
