+++ src/components/ui-components/MultiAttributeSelector.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement, useEffect } from "react";
 import Select from "./Select";
-import { useAttributes } from "../lib/utils/hooks";
+import { useAttributes } from "../../lib/utils/hooks";
 import "./styles/MultiAttributeSelector.css";
 
 interface MultiAttributeSelectorProps {
